Исправления в валидации + дробление кода
