+++ js/data.js
@@ -3,6 +3,7 @@ const LIKE_MIN = 15;
 const LIKE_MAX = 200;
 const COMMENTS_MAX = 30;
 const AVATAR_ID_MAX = 6;
+const ONETIME_ADDED_COMMENTS = 5;
 
 const MESSAGES = [
   'Всё отлично!',
@@ -30,4 +31,4 @@ const DESCRIPTIONS = [
   'Площадка'
 ];
 
-export {MESSAGES, NAMES, DESCRIPTIONS, DESCRIPTION_COUNT, LIKE_MIN, LIKE_MAX, COMMENTS_MAX, AVATAR_ID_MAX};
+export {MESSAGES, NAMES, DESCRIPTIONS, DESCRIPTION_COUNT, LIKE_MIN, LIKE_MAX, COMMENTS_MAX, AVATAR_ID_MAX, ONETIME_ADDED_COMMENTS};
