+++ js/uploading-form.js
@@ -4,7 +4,6 @@ const imgUploadOverlay = uploadingImgForm.querySelector('.img-upload__overlay');
 const imgUploadButtonCancel = uploadingImgForm.querySelector('.img-upload__cancel');
 const userImgComment = uploadingImgForm.querySelector('.text__description');
 const userImgHashtags = uploadingImgForm.querySelector('.text__hashtags');
-const hashtagValid = /^#[a-zа-яё0-9]{1,19}$/i;
 
 const onOverlayKeydown = (evt) => {
   if (evt.key === 'Escape') {
@@ -68,37 +67,4 @@ function onHashtagBlur () {
   userImgHashtags.removeEventListener('blur', onCommentBlur);
 }
 
-const pristine = new Pristine(uploadingImgForm);
-const validateComment = (value) => value.length < 10;
-const validateHashtag = (value) => {
-  const hashtagsArray = value.trim().toLowerCase().split(' ');
-  if (hashtagsArray.length > 5) {
-    return false;
-  }
-  if (hashtagsArray.every((hashtag) => hashtagsArray.filter((item) => hashtag === item).length > 1)) {
-    return false;
-  }
-  if (hashtagsArray.every((hashtag) => hashtagValid.test(hashtag))) {
-    return true;
-  }
-};
-
-pristine.addValidator(
-  userImgComment,
-  validateComment,
-  'Длина комментария не может составлять больше 140 символов'
-);
-
-pristine.addValidator(
-  userImgHashtags,
-  validateHashtag,
-  'Не соответствует критериям'
-);
-
-uploadingImgForm.addEventListener('submit', (evt) => {
-  evt.preventDefault();
-  if (pristine.validate()) {
-    uploadingImgForm.submit();
-  }
-});
-
+export {userImgComment, userImgHashtags, uploadingImgForm};
